requirements.txt
asgiref==3.6.0
Django==4.1.7
gunicorn==20.1.0
PyMySQL==1.0.2
python-dotenv==1.0.0
sqlparse==0.4.3
